start file and start pos

+++ recovery.sh
@@ -134,6 +134,12 @@ if [[ $mysql_start_gtid == '' ]];then
   exit -6
 fi
 echo "mysql_start_gtid=$mysql_start_gtid" >> temp.config
+mysql_gtid_set=`echo "$mysql_recovery_server_stdout"|tail -4|head -1|grep mysql_gtid_set|cut -f 2 -d'='`
+if [[ $mysql_gtid_set == '' ]];then
+  echo "[`date +%Y%m%d%H%M%S`] | fileanme: "$BASH_SOURCE" | line_number: "$LINENO" | Can not find mysql_gtid_set in xbackup backup file"
+  exit -6
+fi
+
 rm -fr $mysql_recovery_backup_temp_dir/*
 
 
@@ -169,6 +175,11 @@ fi
 echo "$mysql_server_stdout"
 
 mysql_end_gtid=`echo "$mysql_server_stdout"|tail -1|grep mysql_end_gtid|cut -f 2 -d'='`
+# mysql_end_gtid需要加1，因为是SQL_BEFORE_GTIDS，需要回放到mysql_end_gtid，还要往后移一位
+mysql_end_gtid_no=`echo "$mysql_end_gtid"|awk -F':' '{print $2}'`
+((mysql_sql_brefore_gtid_no = mysql_end_gtid_no + 1))
+mysql_sql_brefore_gtid=`echo "$server_uuid:$mysql_sql_brefore_gtid_no"`
+mysql_gtid_set=`echo "$mysql_gtid_set"|sed 's/'$server_uuid':[0-9]\{1,\}-[0-9]\{1,\}/'$mysql_sql_brefore_gtid'/g'`
 binlog_package=`echo "$mysql_server_stdout"|tail -2|head -1|grep binlog_package|cut -f 2 -d'='`
 
 scp $mysql_linux_user@$mysql_host:$mysql_backup_temp_dir/$binlog_package $mysql_recovery_backup_temp_dir/
@@ -316,14 +327,9 @@ if [ $? -ne 0 ];then
 fi
  
 ## 只需要开启SQL线程对指定的relay log开始回放即可
-# mysql_end_gtid需要加1，因为是SQL_BEFORE_GTIDS，需要回放到mysql_end_gtid，还要往后移一位
-mysql_end_gtid_no=`echo "$mysql_end_gtid"|awk -F':' '{print $2}'`
-mysql_uuid=`echo "$mysql_end_gtid"|awk -F':' '{print $1}'`
-((mysql_sql_brefore_gtid_no = mysql_end_gtid_no + 1))
-mysql_sql_brefore_gtid=`echo "$mysql_uuid:$mysql_sql_brefore_gtid_no"`
-$recovery_mysql_client -e "START SLAVE SQL_THREAD UNTIL SQL_BEFORE_GTIDS='$mysql_sql_brefore_gtid';"
+$recovery_mysql_client -e "START SLAVE SQL_THREAD UNTIL SQL_BEFORE_GTIDS='$mysql_gtid_set';"
 if [ $? -ne 0 ];then
-  echo "[`date +%Y%m%d%H%M%S`] | fileanme: "$BASH_SOURCE" | line_number: "$LINENO" | start slave sql_thread UNTIL SQL_BEFORE_GTIDS=$mysql_sql_brefore_gtid failed"
+  echo "[`date +%Y%m%d%H%M%S`] | fileanme: "$BASH_SOURCE" | line_number: "$LINENO" | start slave sql_thread UNTIL SQL_BEFORE_GTIDS=$mysql_gtid_set failed"
   exit -1
 fi
